Move ViewPagerNonSwipeable.kt to pandautils so it can be used in the teacher app. Removed redundant ViewPagerNoSwipe class.
