+++ apps/student/src/main/res/layout/course_module_progression.xml
@@ -23,7 +23,7 @@
     android:orientation="vertical"
     android:id="@+id/moduleProgressionPage">
 
-    <com.instructure.student.view.ViewPagerNonSwipeable
+    <com.instructure.pandautils.views.ViewPagerNonSwipeable
         android:id="@+id/viewPager"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
