+++ apps/student/src/main/res/layout/fragment_syllabus.xml
@@ -67,7 +67,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <com.instructure.student.view.ViewPagerNonSwipeable
+        <com.instructure.pandautils.views.ViewPagerNonSwipeable
             android:id="@+id/syllabusPager"
             android:layout_width="match_parent"
             android:layout_height="match_parent" />
