+++ apps/teacher/src/main/res/layout/fragment_syllabus.xml
@@ -1,5 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
+<?xml version="1.0" encoding="utf-8"?><!--
   ~ Copyright (C) 2019 - present Instructure, Inc.
   ~
   ~     This program is free software: you can redistribute it and/or modify
@@ -67,8 +66,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <!-- TODO Add Non Swipable ViewPager? -->
-        <androidx.viewpager.widget.ViewPager
+        <com.instructure.pandautils.views.ViewPagerNonSwipeable
             android:id="@+id/syllabusPager"
             android:layout_width="match_parent"
             android:layout_height="match_parent" />
