+++ libs/pandautils/src/main/java/com/instructure/pandautils/views/ViewPagerNonSwipeable.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017 - present  Instructure, Inc.
+ * Copyright (C) 2018 - present Instructure, Inc.
  *
  *     This program is free software: you can redistribute it and/or modify
  *     it under the terms of the GNU General Public License as published by
@@ -12,23 +12,28 @@
  *
  *     You should have received a copy of the GNU General Public License
  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
  */
-package com.instructure.teacher.view
+
+package com.instructure.pandautils.views
 
 import android.content.Context
-import androidx.viewpager.widget.ViewPager
 import android.util.AttributeSet
 import android.view.MotionEvent
+import androidx.viewpager.widget.ViewPager
 
-class ViewPagerNoSwipe : ViewPager {
+class ViewPagerNonSwipeable : ViewPager {
+
+    var canSwipe = false
 
     constructor(context: Context) : super(context)
-
     constructor(context: Context, attrs: AttributeSet) : super(context, attrs)
 
-    override fun onInterceptTouchEvent(ev: MotionEvent): Boolean = false
+    // Never allow swiping to switch between pages
+    override fun onInterceptTouchEvent(event: MotionEvent): Boolean = if (canSwipe) super.onInterceptTouchEvent(event) else false
 
-    override fun onTouchEvent(ev: MotionEvent): Boolean = false
+    // Never allow swiping to switch between pages
+    override fun onTouchEvent(event: MotionEvent): Boolean = if (canSwipe) super.onTouchEvent(event) else false
 
-    override fun canScrollHorizontally(direction: Int): Boolean = false
+    override fun canScrollHorizontally(direction: Int): Boolean = if (canSwipe) super.canScrollHorizontally(direction) else false
 }
